+++ changes.md
@@ -3,6 +3,7 @@
 0.11 and 0.12 bring a number of small breaking changes that were necessary to fix some flaws in Onyx's usability. After 0.12 the pace of these breaking changes will be reduced.
 
 * New api function `onyx.api/job-ids-history`. Allows reverse lookup of a :job-name job-key, to its corresponding history of job-id and tenancy-ids. This makes supporting long running migrateable jobs much easier.
+* New api function `onyx.api/job-state`. Plays back the log for a given tenancy-id and job-id and returns the current state of the job.
 * `onyx.api/job-snapshot-coordinates` was improved by use of `onyx.api/job-ids-history`, and will now walk backwards through the job history until it finds an available snapshot.
 * onyx-http has been restored to functionality. Thanks [Alexander Solovyov](https://github.com/piranha)
 * **BREAKING CHANGE** Flow condition handling of exceptions has been altered. `:flow/to` tasks that receive exception messages will no longer receive non-exceptional messages, ensuring that simple error handling flows are possible.